requirements.txt
scipy>=1.2.1
numpy>=1.16.1
django-crispy-forms>=1.7.2
matplotlib>=3.0.2
PyYAML>=5.1
python-socketio>=5.0.0
websocket-client>=1.0.0
dataclasses; python_version < "3.7"
